feat: Update useParsedMutation to support different HTTP methods
This commit modifies the useParsedMutation composable to support different HTTP methods for making parsed mutations. It introduces a new options parameter with a default value of { method: 'POST' } and adds a getMutator function to determine the appropriate axios instance method based on the specified method. The mutator function is then used to make the mutation request.

The purpose of this change is to enhance the flexibility of the useParsedMutation composable by allowing developers to specify the HTTP method for their mutations.