feat: Add PUT and PATCH functionality to HomeView
This commit adds PUT and PATCH functionality to the HomeView component. It introduces two new buttons, "Put Post" and "Patch Post", which trigger the corresponding mutation requests using the `putMutation` and `patchMutation` composable functions. The buttons are disabled when the mutations are pending.

The purpose of this change is to allow users to update existing posts by sending PUT and PATCH requests to the server.