feat: Add delete functionality to HomeView
This commit adds a new button, "Delete Post", to the HomeView component. Clicking on this button triggers the `deletePost` function, which sends a DELETE request to the server using the `deleteMutation` composable function. The button is disabled when the delete mutation is pending.

The purpose of this change is to allow users to delete posts by sending a DELETE request to the server.