feat: Refactor addPost, putPost, and deletePost functions in HomeView
This commit refactors the addPost, putPost, and deletePost functions in the HomeView component to pass the post ID as a parameter. Instead of hardcoding the post ID in the URL, the functions now generate a random post ID between 1 and 100 using Math.random(). This change improves the reusability and flexibility of the functions.

The purpose of this change is to make the add, update, and delete functionality in the HomeView component more dynamic and scalable.